requirements.txt
# Minimal requirements for improved video analysis
# Focused on Google's video understanding capabilities

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# AI APIs - Core requirements only
openai==1.6.1
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0

# Optional: For fallback video info (if needed)
pytubefix==6.10.2

# Note: Removed heavy dependencies:
# - torch/transformers (not needed with Gemini)
# - sentence-transformers (using OpenAI embeddings)
# - opencv-python (not needed for direct video analysis)
# - moviepy (not needed with Gemini video API)
# - scikit-learn (simplified approach)